# schema/openapi.yaml
...
openapi: 3.0.3
info:
  title: Tesserae REST API
  version: 0.72.0
  description: |
    _Self-hosted e-ink dashboard server, device + automation HTTP API._


    Tesserae is a self-hosted hobby project that composes tile-based
    e-ink dashboards, renders them headless, and pushes the frames to
    panel devices.

    This spec covers four surfaces an external client touches:

    1. **Native REST device API** under `/api/v1/device/*` for clients
       that prefer HTTP polling over MQTT. Per-device bearer-token
       auth bootstrapped through pairing codes or MAC-based discover.
    2. **TRMNL-compatible BYOS** (`/api/display`, `/api/setup`,
       `/api/log*`) for the Terminus / TRMNL native firmware family.
    3. **Webhook push** (`/api/v1/push`, `/api/v1/push/image`) for
       automations (Home Assistant beyond MQTT, n8n, cron + curl, phone
       shortcuts) that trigger a render + publish of a saved page, or
       send a single image straight to named displays.
    4. **Static render artifacts** (`/renders/*`, `/preview/*.png`,
       `/mirror/<id>`) that clients fetch after being told their URL
       through one of the above.

    A health probe (`/healthz`) is included for monitoring.
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Tesserae
    url: https://github.com/dmellok/tesserae
servers:
  - url: "{scheme}://{host}"
    description: A self-hosted Tesserae instance.
    variables:
      scheme:
...
security:
  - DeviceToken: []
paths:
...
  /api/v1/push:
    post:
      tags: [webhook]
      summary: Trigger a render + publish for a saved page.
      description: |
        External automation hook. Authenticated by the global
        webhook token from Settings → System → Webhook (sent as either
        `Authorization: Bearer …` or `X-Tesserae-Token: …`).

        `device_ids` is optional. When omitted, the push fans out to
        every device the page is bound to. Per-device quiet hours
        are respected; a fully-quiet fan-out returns 202 with
        `status: "quiet"` so the caller can tell that apart from a
        real failure.

        Status code map:

        | `status`     | HTTP |
        | ------------ | ---- |
        | `sent`       | 200  |
        | `quiet`      | 202  |
        | `not_found`  | 404  |
        | `busy`       | 409  |
        | `failed`     | 500  |
      operationId: postWebhookPush
      security:
        - WebhookBearer: []
        - WebhookToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPushRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/WebhookPushRequest'
      responses:
        '200':
          description: Page rendered + published.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookPushResponse'
        '202':
          description: All bound devices are in quiet hours; nothing was sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookPushResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Page id not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookPushResponse'
        '409':
          description: Another push for this page is already in flight.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookPushResponse'
        '500':
          description: Render or publish failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookPushResponse'
        '503':
          description: Webhook token not generated yet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookPushResponse'
  /api/v1/push/image:
    post:
      tags: [webhook]
      summary: Push a single image straight to one or more displays.
      description: |
        For when there is no saved dashboard behind the frame: a photo, a
        chart another tool generated, an already-rendered image. Same
        global webhook token as `POST /api/v1/push`.

        Unlike `/api/v1/push`, `device_ids` is **required**: a dashboard
        knows which displays it belongs to, a loose image does not, and
        defaulting to every registered panel is the wrong way to find out.
        Send it comma-separated or as a repeated field.

        The image is fitted and dithered to each target's panel exactly as
        the Send page does. Accepted types: JPEG, PNG, HEIC, HEIF, WebP,
        up to 25 MiB encoded and 8192 px on the longest edge. EXIF
        orientation is applied, so a phone photo arrives the way its
        camera roll shows it; `rotate` turns it further when the image and
        the panel disagree about which way is up.

        Per-device quiet hours are respected unless `override_quiet_hours`
        is set. The response lists each display under `sent`, `quiet`, or
        `failed`, so a partial result is readable without inferring it
        from a count.

        Status code map:

        | `status`  | HTTP | Meaning                                  |
        | --------- | ---- | ---------------------------------------- |
        | `sent`    | 200  | at least one display took the frame      |
        | `quiet`   | 202  | every target is inside its quiet hours   |
        | `failed`  | 500  | at least one display failed              |
      operationId: postWebhookPushImage
      security:
        - WebhookBearer: []
        - WebhookToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image, device_ids]
              properties:
                image:
                  type: string
                  format: binary
                  description: The image file.
                device_ids:
                  type: string
                  description: >-
                    Target display ids, comma-separated or repeated.
                  example: pi_kitchen,pi_bedroom
                fit:
                  type: string
                  enum: [fit, fill, blur, stretch, center]
                  description: >-
                    How to place a non-panel-sized image. Defaults to the
                    renderer's own default when omitted.
                rotate:
                  type: string
                  enum: [auto, "0", "90", "180", "270"]
                  description: >-
                    Turn the image clockwise before fitting it, for when
                    the image's orientation and the panel's disagree.
                    `auto` turns a quarter only when their aspects are
                    opposite, so a portrait photo fills a landscape panel
                    instead of landing as a letterboxed strip. Omitted,
                    the image keeps the orientation it was shot in, which
                    is what an image carrying text wants. EXIF orientation
                    is applied either way.
                label:
                  type: string
                  description: History label for the push.
                  default: Webhook image
                override_quiet_hours:
                  type: boolean
                  default: false
                  description: >-
                    Send even to displays currently inside their quiet
                    hours. Off by default; set it for images that are worth
                    a refresh at 3am (a doorbell snapshot, an alarm frame)
                    rather than for ambient content.
      responses:
        '200':
          description: The image was published to at least one display.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookImagePushResponse'
        '202':
          description: Every target display is in quiet hours; nothing was sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookImagePushResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: One or more display ids are unknown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookImagePushResponse'
        '413':
          description: The image exceeds the size or edge-length limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookImagePushResponse'
        '415':
          description: Unsupported or undecodable image type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookImagePushResponse'
        '500':
          description: One or more displays failed to take the frame.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookImagePushResponse'
        '503':
          description: Webhook token not generated yet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookImagePushResponse'

  # ------------------------------------------------------------------
  # Frame artifacts
  # ------------------------------------------------------------------
...
components:
  securitySchemes:
    DeviceToken:
      type: http
      scheme: bearer
      bearerFormat: hex
      description: |
        Per-device access token issued by `/api/v1/device/register`
        or `/api/v1/device/discover`. Sent as
        `Authorization: Bearer <token>`.
    PairingCode:
      type: apiKey
      in: header
      name: X-Pairing-Code
      description: |
        6-digit code (`[0-9]{6}`) generated by the admin under
        Settings → Devices → Pair. Single-use, 15-minute TTL.
    TrmnlAccessToken:
      type: apiKey
      in: header
      name: access-token
      description: |
        TRMNL BYOS legacy auth header. Equivalent to the `Authorization`
        bearer header; either is accepted on the `/api/*` BYOS
        endpoints.
    TrmnlAuthBearer:
      type: http
      scheme: bearer
      description: TRMNL BYOS bearer-style auth header.
    WebhookBearer:
      type: http
      scheme: bearer
      description: |
        Global webhook token from Settings → System → Webhook.
        Single token shared by all webhook callers (HA, n8n, cron, …).
    WebhookToken:
      type: apiKey
      in: header
      name: X-Tesserae-Token
      description: |
        Alternate header for the global webhook token, for callers
        whose tooling can't customise the `Authorization` header.
...
  responses:
    BadRequest:
      description: The request body failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid credential.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [status, error]
      properties:
        status:
          type: integer
          description: HTTP status code echoed in the body for convenience.
        error:
          type: string
          description: Human-readable explanation.
...
    WebhookPushRequest:
      type: object
      required: [page_id]
      properties:
        page_id:
          type: string
          description: Id of the saved page to render + publish.
          example: hallway
        device_ids:
          oneOf:
            - type: array
              items:
                type: string
            - type: string
              description: Comma-separated form-encoded fallback.
          description: |
            Subset of device ids to push to. Omit to fan out to every
            device the page is bound to.
          example: [pi_kitchen, pi_bedroom]
    WebhookPushResponse:
      type: object
      required: [status, page_id]
      properties:
        status:
          type: string
          enum: [sent, quiet, busy, not_found, failed, unauthorized, bad_request, disabled]
        page_id:
          type: string
        duration_s:
          type: number
          format: float
          nullable: true
          description: Wall-clock time for the render + publish.
        event_id:
          type: string
          nullable: true
          description: Id of the row written to the Events log (when applicable).
        error:
          type: string
          nullable: true
          description: Failure detail when `status` is `failed` / `bad_request` / `not_found`.
    WebhookImagePushResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [sent, quiet, failed, unauthorized, bad_request, not_found, disabled]
        sent:
          type: array
          description: Displays that took the frame.
          items:
            type: string
        quiet:
          type: array
          description: Displays skipped because they are inside their quiet hours.
          items:
            type: string
        failed:
          type: array
          description: Displays that could not be reached, with the reason for each.
          items:
            type: object
            properties:
              device_id:
                type: string
              error:
                type: string
        error:
          type: string
          nullable: true
          description: Failure detail when the request itself was rejected.